Move context types to steps package
