+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/BuildCacheStep.java
@@ -28,9 +28,7 @@ import org.gradle.caching.internal.origin.OriginMetadata;
 import org.gradle.internal.Try;
 import org.gradle.internal.execution.CurrentSnapshotResult;
 import org.gradle.internal.execution.ExecutionOutcome;
-import org.gradle.internal.execution.IncrementalChangesContext;
 import org.gradle.internal.execution.OutputChangeListener;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.caching.CachingState;
 import org.gradle.internal.file.Deleter;
