+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/BuildOperationStep.java
@@ -16,9 +16,7 @@
 
 package org.gradle.internal.execution.steps;
 
-import org.gradle.internal.execution.Context;
 import org.gradle.internal.execution.Result;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.operations.BuildOperationContext;
 import org.gradle.internal.operations.BuildOperationDescriptor;
 import org.gradle.internal.operations.BuildOperationExecutor;
