+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/CancelExecutionStep.java
@@ -18,9 +18,7 @@ package org.gradle.internal.execution.steps;
 
 import org.gradle.api.BuildCancelledException;
 import org.gradle.initialization.BuildCancellationToken;
-import org.gradle.internal.execution.Context;
 import org.gradle.internal.execution.Result;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.execution.UnitOfWork;
 
 public class CancelExecutionStep<C extends Context> implements Step<C, Result> {
