+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/Context.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution;
+package org.gradle.internal.execution.steps;
 
-import org.gradle.internal.execution.history.changes.ExecutionStateChanges;
-
-import java.util.Optional;
-
-public interface IncrementalChangesContext extends CachingContext {
-    /**
-     * Returns changes detected between the execution state after the last execution and before the current execution.
-     * Empty if changes couldn't be detected (e.g. because history was unavailable).
-     */
-    Optional<ExecutionStateChanges> getChanges();
+public interface Context {
 }
