+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/DeferredExecutionAwareStep.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution;
+package org.gradle.internal.execution.steps;
 
-import org.gradle.internal.execution.history.ExecutionHistoryStore;
+import org.gradle.cache.Cache;
+import org.gradle.internal.Try;
+import org.gradle.internal.execution.DeferredExecutionHandler;
+import org.gradle.internal.execution.Result;
+import org.gradle.internal.execution.UnitOfWork;
+import org.gradle.internal.execution.UnitOfWork.Identity;
 
-import java.io.File;
-import java.util.Optional;
-
-public interface WorkspaceContext extends IdentityContext {
-    File getWorkspace();
-
-    Optional<ExecutionHistoryStore> getHistory();
+public interface DeferredExecutionAwareStep<C extends Context, R extends Result> extends Step<C, R> {
+    <T, O> T executeDeferred(UnitOfWork work, C context, Cache<Identity, Try<O>> cache, DeferredExecutionHandler<O, T> handler);
 }
