+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/ExecutionRequestContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution;
+package org.gradle.internal.execution.steps;
 
-import org.gradle.cache.Cache;
-import org.gradle.internal.Try;
-import org.gradle.internal.execution.UnitOfWork.Identity;
+import java.util.Optional;
 
-public interface DeferredExecutionAwareStep<C extends Context, R extends Result> extends Step<C, R> {
-    <T, O> T executeDeferred(UnitOfWork work, C context, Cache<Identity, Try<O>> cache, DeferredExecutionHandler<O, T> handler);
+public interface ExecutionRequestContext extends Context {
+    /**
+     * If incremental mode is disabled, this returns the reason, otherwise it's empty.
+     */
+    Optional<String> getRebuildReason();
 }
