+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/IdentityCacheStep.java
@@ -19,11 +19,8 @@ package org.gradle.internal.execution.steps;
 import org.gradle.cache.Cache;
 import org.gradle.internal.Cast;
 import org.gradle.internal.Try;
-import org.gradle.internal.execution.DeferredExecutionAwareStep;
 import org.gradle.internal.execution.DeferredExecutionHandler;
-import org.gradle.internal.execution.IdentityContext;
 import org.gradle.internal.execution.Result;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.UnitOfWork.Identity;
