+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/InputChangesContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,13 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution;
+package org.gradle.internal.execution.steps;
 
-public interface Context {
+import org.gradle.internal.execution.history.changes.InputChangesInternal;
+
+import java.util.Optional;
+
+public interface InputChangesContext extends BeforeExecutionContext {
+    Optional<InputChangesInternal> getInputChanges();
+    boolean isIncrementalExecution();
 }
