+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/RecordOutputsStep.java
@@ -16,9 +16,7 @@
 
 package org.gradle.internal.execution.steps;
 
-import org.gradle.internal.execution.Context;
 import org.gradle.internal.execution.CurrentSnapshotResult;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.history.OutputFilesRepository;
