+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/Step.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution;
+package org.gradle.internal.execution.steps;
 
-import java.util.Optional;
+import org.gradle.internal.execution.Result;
+import org.gradle.internal.execution.UnitOfWork;
 
-public interface ExecutionRequestContext extends Context {
-    /**
-     * If incremental mode is disabled, this returns the reason, otherwise it's empty.
-     */
-    Optional<String> getRebuildReason();
+public interface Step<C extends Context, R extends Result> {
+    R execute(UnitOfWork work, C context);
 }
