+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/TimeoutStep.java
@@ -18,9 +18,7 @@ package org.gradle.internal.execution.steps;
 
 import org.gradle.api.GradleException;
 import org.gradle.api.InvalidUserDataException;
-import org.gradle.internal.execution.Context;
 import org.gradle.internal.execution.Result;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.timeout.Timeout;
 import org.gradle.internal.execution.timeout.TimeoutHandler;
