+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/legacy/MarkSnapshottingInputsFinishedStep.java
@@ -16,10 +16,10 @@
 
 package org.gradle.internal.execution.steps.legacy;
 
-import org.gradle.internal.execution.CachingContext;
 import org.gradle.internal.execution.Result;
-import org.gradle.internal.execution.Step;
 import org.gradle.internal.execution.UnitOfWork;
+import org.gradle.internal.execution.steps.CachingContext;
+import org.gradle.internal.execution.steps.Step;
 
 /**
  * This is a temporary measure for Gradle tasks to track a legacy measurement of all input snapshotting together.
