+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/AssignWorkspaceStepTest.groovy
@@ -16,10 +16,9 @@
 
 package org.gradle.internal.execution.steps
 
-import org.gradle.internal.execution.IdentityContext
+
 import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
-import org.gradle.internal.execution.WorkspaceContext
 import org.gradle.internal.execution.workspace.WorkspaceProvider
 
 class AssignWorkspaceStepTest extends StepSpec<IdentityContext> {
