+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/BroadcastChangingOutputsStepTest.groovy
@@ -20,7 +20,6 @@ import org.gradle.api.file.FileCollection
 import org.gradle.internal.execution.OutputChangeListener
 import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
-import org.gradle.internal.execution.WorkspaceContext
 import org.gradle.internal.file.TreeType
 
 class BroadcastChangingOutputsStepTest extends StepSpec<WorkspaceContext> {
