+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/CreateOutputsStepTest.groovy
@@ -19,7 +19,6 @@ package org.gradle.internal.execution.steps
 import org.gradle.api.internal.file.TestFiles
 import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
-import org.gradle.internal.execution.WorkspaceContext
 import org.gradle.internal.file.TreeType
 
 class CreateOutputsStepTest extends StepSpec<WorkspaceContext> {
