+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/IdentityCacheStepTest.groovy
@@ -21,7 +21,6 @@ import org.gradle.cache.ManualEvictionInMemoryCache
 import org.gradle.internal.Try
 import org.gradle.internal.execution.CachingResult
 import org.gradle.internal.execution.DeferredExecutionHandler
-import org.gradle.internal.execution.IdentityContext
 import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
