+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/ResolveCachingStateStepTest.groovy
@@ -17,8 +17,6 @@
 package org.gradle.internal.execution.steps
 
 import org.gradle.caching.internal.controller.BuildCacheController
-import org.gradle.internal.execution.BeforeExecutionContext
-import org.gradle.internal.execution.CachingContext
 import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.execution.caching.CachingDisabledReason
 import org.gradle.internal.execution.caching.CachingDisabledReasonCategory
