+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/ResolveInputChangesStepTest.groovy
@@ -16,8 +16,7 @@
 
 package org.gradle.internal.execution.steps
 
-import org.gradle.internal.execution.IncrementalChangesContext
-import org.gradle.internal.execution.InputChangesContext
+
 import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.execution.history.changes.ExecutionStateChanges
