+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/StepSpec.groovy
@@ -16,8 +16,7 @@
 
 package org.gradle.internal.execution.steps
 
-import org.gradle.internal.execution.Context
-import org.gradle.internal.execution.DeferredExecutionAwareStep
+
 import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.operations.BuildOperationType
 import org.gradle.internal.operations.TestBuildOperationExecutor
